% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/delete_food_plan.R
\name{delete_food_plan}
\alias{delete_food_plan}
\title{Delete food plan}
\usage{
delete_food_plan()
}
\arguments{
\item{token}{token, output from drop_auth_refreshable}
}
\description{
Delete food plan
}
