% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/download_food_plan.R
\name{download_food_plan}
\alias{download_food_plan}
\title{Download food plan}
\usage{
download_food_plan(data, file, token)
}
\arguments{
\item{data}{tibble with food plan}

\item{file}{character string}

\item{token}{token, output from drop_auth_refreshable}
}
\description{
Download food plan
}
