% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/drop_auth_refreshable.R
\name{drop_auth_refreshable}
\alias{drop_auth_refreshable}
\title{Authenticate dropbox with refreshable tokens}
\usage{
drop_auth_refreshable(
  new_user = FALSE,
  key = "mmhfsybffdom42w",
  secret = "l8zeqqqgm1ne5z0",
  cache = TRUE,
  rdstoken = NA
)
}
\description{
Authenticate dropbox with refreshable tokens
}
\details{
This has been placed in the app_server.R:
.dstate <<- new.env(parent = emptyenv())
}
