% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/init_food_plan.R
\name{init_food_plan}
\alias{init_food_plan}
\title{Initialize food plan}
\usage{
init_food_plan(token)
}
\arguments{
\item{token}{token, output from drop_auth_refreshable}
}
\description{
Initialize food plan
}
