% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/load_food_plan.R
\name{load_food_plan}
\alias{load_food_plan}
\title{Load food plan}
\usage{
load_food_plan(token)
}
\arguments{
\item{token}{token, output from drop_auth_refreshable}
}
\value{
tibble with existing food plan (empty food plan if non-existing)
}
\description{
Load food plan
}
