% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pull_subset.R
\name{pull_subset}
\alias{pull_subset}
\title{Pull subset}
\usage{
pull_subset(data, var, ...)
}
\arguments{
\item{data}{tibble (see `load_food_plan`)}

\item{var}{character string with variable name}

\item{...}{additional filtering}
}
\description{
Pull subset
}
