% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/replace_eol.R
\name{replace_eol}
\alias{replace_eol}
\title{Replace end of line}
\usage{
replace_eol(x, old = "\\\\n", new = "--break--")
}
\arguments{
\item{x}{character vector}

\item{old}{old eol (end of line) character (defalt is \n)}

\item{new}{new eol (end of line) character (default is --break--)}
}
\description{
Replace end of line
}
\details{
`replace_eol` allows line breaking characters in .txt files without having issues when loading/saving tables
}
