% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/save_food_plan.R
\name{save_food_plan}
\alias{save_food_plan}
\title{Save food plan}
\usage{
save_food_plan(date, breakfast, lunch, dinner, token)
}
\arguments{
\item{date}{date in the format yyyy-mm-dd}

\item{breakfast}{character string with information on breaskfast}

\item{lunch}{character string with information on lunch}

\item{dinner}{character string with information on dinner}

\item{token}{token, output from drop_auth_refreshable}
}
\value{
named list with:
* df_input: tibble with all inputs
* df_saved: tibble with inputs modified according to `replace_eol`
}
\description{
Save food plan
}
