% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/save_food_plan.R
\name{save_food_plan}
\alias{save_food_plan}
\title{Save food plan}
\usage{
save_food_plan(date, breakfast, lunch, dinner)
}
\arguments{
\item{date}{date in the format yyyy-mm-dd}

\item{breakfast}{character string with information on breaskfast}

\item{lunch}{character string with information on lunch}

\item{dinner}{character string with information on dinner}
}
\value{
invisible TRUE
}
\description{
Save food plan
}
